helm: set default value for cluster.id option
[ upstream commit a0e0917c23dcf3e4fa4709b88f7cfb0e8c5ae6b0 ]

When using the ConfigMap, the `cluster-id` option is loaded from the
relevant file. Without the `0`, that file is empty, so Viper (silently)
fails to convert an empty string to an int.

Even if the cluster ID option is left with a 0 (the Go zero value for
the int type), we specify the correct default value to avoid a silent
failure while loading the option value.

[ Backport notes: Trivial conflict, solved by regenerating
    install/kubernetes/cilium/README.md. ]

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>