bpf: Provision HostPort also for case of Maglev
[ upstream commit cc93f32ad72612a6699950934cec8d898d1b279b ]

Previosly, added HostPort service will not create a map in Maglev LUT, so
access to HostPort from outside of the host will cause service backend not
found error and thus drop the request.

PR #19882 has added a quirk into the fast-path to skip Maglev LUT lookup,
but better approach would be to extend useMaglev() agent code and if needed
we could optimize the size of such maps. Expectation here anyway is that
there may not be too many HostPort services provisioned in general.

Fixes: #19260
Reported-by: Rel Zhong <relzhong@outlook.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Link: https://github.com/cilium/cilium/pull/19882
Signed-off-by: Quentin Monnet <quentin@isovalent.com>