Move GRPCRoute into refactored Gateway reconciler
[ upstream commit 23b4cc0 ]

This commit moves GRPCRoute into the refactored Gateway reconciler.

It also fixes an existing issue with GRPCRoute reconciliation
(#39021 and #40922), by adding correct `supportedKinds` handling.

While fixing this, also fixed some inconsistencies in Condition handling
across all Gateway API objects, so this is a much bigger commit than
expected.

`Programmed` is now added as `Pending` before addresses are added to the
Gateway, and is updated to `Programmed` when addresses are added.

Tests have been updated to reflect this.

Ordering of Conditions in status has also been made more consistent.

The PEM checking performed in `isValidPEMFormat` has been loosened,
and this function will now return `true` for all byte slices that
contain at least one valid PEM-encoded certificate or key. Previous
behavior was that it would only return if _all_ bytes decoded
successfully, which can cause issues if there are extra line feeds or
extraneous material in the certificate - that don't matter to Envoy
or to the eventual functionality.

Co-authored-by: Archonr <archonr94@gmail.com>
Co-authored-by: Vlad Korkach <v.korkach@playson.com>
Signed-off-by: Nick Young <nick@isovalent.com>