Remove redundant ReferenceGrant reconciler
[ upstream commit e098b30 ]

This commit removes the now-redundant ReferenceGrant
reconciler, which was a placeholder in case we
needed to update ReferenceGrant status at a later
date.

If we do need to do that in the future, which is unlikely,
we can use similar mechanisms to what we now use for
Route status in the Gateway reconciler.

Signed-off-by: Nick Young <nick@isovalent.com>