Add test for Programmed Condition behavior
[ upstream commit 4c81d83 ]

As part of the previous refactor, the Programmed
Condition now is correctly set to `Pending` when
there is no address set on the Gateway.

This adds a test that, when there is an address
set on the Gateway, the Programmed Condition
is correctly updated.

The address is set by taking it from the generated
Loadbalancer Service, so this test supplies a
preconfigured Loadbalancer Service to ensure an
address gets added to the Gateway correctly.

Other changes as a result of review have also
been incorporated.

Signed-off-by: Nick Young <nick@isovalent.com>