refactor: move gateway auto discover to a new reconciler
[ upstream commit a07066dcae158cc0a3ef684a0fac0f01580e1020 ]
[ backporter's note: adapted setupStateDB to older StateDB version ]

This commit fixes an issue with bgp gateway auto discovery.
Previously there was an issue with the priority of the reconcilers
where the gateway ip set by the neighbor reconciler auto discovery
is not picked up by the pod cidr reconciler when setting the
policy.
This fix refactors the auto discovery logic to a new reconciler that
runs first.

Fixes: #41904

Signed-off-by: Jiasheng Zhu <gnehsaijuhz@hotmail.com>
Signed-off-by: Jussi Maki <jussi@isovalent.com>