bpf: port replace before l4 checksum replace
[ upstream commit 33e11da5c8791ae3bda1af406fbcdff3c3713e72 ]

When rewriting the inner header within
the payload of an ICMP error packet,
the l4_modify_port() function is also used.

In such cases, the inner L4 header may be truncated
and not include the checksum field.

When that happens, the function returns DROP_CSUM_L4,
but the caller handles this error and continues processing.
https://github.com/cilium/cilium/commit/3c3f80bd48bc

Before returning DROP_CSUM_L4,
the port in the inner L4 header should be modified firstly.

Signed-off-by: YushoYamaguchi <ysh.824@outlook.jp>
Signed-off-by: Jussi Maki <jussi@isovalent.com>