.github/workflows: Add base-SHA input to ariane triggered workflows
For certain workflows we would like to be able to do comparisons between
the PR head SHA and the base branch SHA. For example to determine
performance regressions.

Currently ariane gives us the `context-ref`, but this is only equal to
the base branch when the PR is opened from a fork. Adding the `base-SHA`
input will allow us to use an exact SHA, and work on all types of PRs.

For now this commit adds the input as non required since we still run
on an older version of ariane that does not provide this input. Once
we have upgraded we can make this required. Adding the input must be
done in the workflows first, since Github does not allow calling a
workflow with an input that is undefined.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>