test: handle race condition in TestNodeManagerAbortReleaseIPReassignment
[ upstream commit 23d84c8501eb289501b73b10ac46abe2ca8ad92a ]

[ backporter's notes: added MaintainIPPool call from upstream,
  which was not present in the target banch test code ]

The node manager test TestNodeManagerAbortReleaseIPReassignment was
flaky in part due to a race condition between the synchronous
MaintainIPPool call in the test code and the background trigger to
periodically call MaintainIPPool in the node manager.

This fix handles the race condition by checking if the returned error is
the “IP not found” error returned by nodeOperationsMock test object.
This is the expected outcome if the MaintainIPPool trigger wins the
race.

Signed-off-by: Liam Connery <lconnery@google.com>
Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>