fqdn: Fix fqdn subsystem correctness issues causing packet drops
[ upstream commit 8895a7b7b5a820977cba7d04200430d03bf22af0 ]

To ensure we never leak/fail to upsert identities, we ensure we only add
IPs that were already present in the local cache.

The error we are fixing here is the one where the domain is already in
the FQDN cache, AND a lookup is in flight. If its added to local
endpoint cache before the GC ReplaceFromCacheByNames is executed, it
could result in it being added to the global cache by the GC,
hence it never being upserted by the dns lookup code. This leads to a
situation where the toFQDN won't allow egress to the given IP - and the
only way to recover for the given IP is to wait until its removed from
the fqdn cache and a new lookup returns the same IP. This is however not
very likely, since the same domain can continue to return the same IP
forever, keeping it constantly in the fqdn cache.

The same partially applies to endpoint created during the GC iteration,
resulting in the GC evicting some of the IPs upserted to the global
cache by that endpoint. That situation could also result in a leak where
its removed from the fqdn cache, but still be leaked in the ipcache.

This still needs extensive testing to ensure it doesn't uncover other
issues that rely on the GC upserting new IPs into the global cache - as
those could now become even more problematic. We have been running a
version of this in v1.15 and v1.16 with great success, and without this
patch we see loads of issue and packet drops due to this issue. We have
not been testing this in high scale in newer versions than those
mentioned, and due to other changes in the fqdn subsystem, we need to
sanity check this thoroughly.

This is still in works, and will be split into multiple commits, and
potentially multiple PRs.

Signed-off-by: Odin Ugedal <odin@uged.al>
Signed-off-by: Odin Ugedal <ougedal@palantir.com>
Signed-off-by: Tobias Klauser <tobias@cilium.io>