bpf: lxc: transfer source identity for lxc-to-host policy
[ upstream commit 324ad59d9aecda82a355ba4ae75f16f9751efc33 ]

[ backporter's notes: only take the parts in bpf_lxc, so that the n+1
  release can trust that all in-flight packets have a valid identity
  prior to the upgrade. ]

Provide the cil_host_policy tailcall program in bpf_host with the
endpoint's actual source identity, so that we can use it for trace/drop
notifications. And with a bit of additional work the HostFW code then can
skip its ipcache lookup to resolve the source identity.

Since this is a cross-program tailcall and packets might be in-flight, we
can't put this identity to use right away.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>