bpf: lxc: always set identity mark on forwarded egressing traffic
[ upstream commit 26602426a0732a65ec597742e6074c939d18326a ]

[ backporter's note: resolved conflicts around added / removed
  code blocks caused by different surrounding context ]

Guessing which exact forward paths can benefit from the identity mark is
fragile, and it's hard to overlook corner cases. Even when we BPF-redirect
a packet to an external interface (with BPF Host Routing) it makes sense
to have the mark, so that the cil_to_netdev can retrieve it without
relying on endpoint map or ipcache.

So let's always set the mark at the very start of the forwarding path.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>
Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>