helm: respect CA bundle configuration in kvstoremesh remote certificate
[ upstream commit c9a594d4e0dbc99f620e9318f9c34ccd1359a385 ]

The blamed commit added the `clustermesh-apiserver-remote-cert` secret
as a projected volume to the clustermesh-apiserver pod, providing the
client certificate and key normally used to authenticate with the etcd
instance in remote clusters, as well as the CA used to validate the
server certificate presented by the etcd instances. However, it missed
respecting the CA bundle setting, incorrectly always using the CA from
the secret itself. Let's get this fixed, aligning the behavior here
with that of all the other clustermesh related secrets.

Fixes: bd5386019c2b ("kvstoremesh: add helm configuration")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Jussi Maki <jussi@isovalent.com>