allow missing verbs for cilium-agent clusterrole
[ upstream commit e12c1d2e4f75f28afeaa74508bab3459b92ffc1b ]

Chart documentation describes that cilium-agent should have GET,LIST,
and WATCH permissions for the secrets resource in default apiGroup when
readSecretsOnlyFromSecretsNamespace is false. Currently, only GET
permission is provided in the role RBAC configuration.

This commit moves the conditional logic of adding the secrets
resource to the apiGroup already defined that has these verbs allowed for
resources in same apiGroup.

Fixes: #42789

Signed-off-by: Tomi Juntunen <tomi.juntunen@iki.fi>
Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>