redirectpolicy: Avoid recomputation when backends do not change
[ upstream commit 64f16d0de0050b2046337756059787076145b958 ]

The comparison of the old and new BackendParams derived from the pods
compared all the fields to see if the backends changed or not to avoid
recomputation. This always failed as the LB control-plane set the 'Source'
field.

Fix this by using a custom comparison that is coupled to how the
BackendParams is produced.

The 'testdata/avoid-recompute.txtar' verifies the fix by comparing
the frontend 'Status.ID' field that it remains the same. The issue
was reproduced with the same test without the fix.

Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>