endpoint: ensure all identity references are properly cleaned up
[ upstream commit f476c332b3bf98a660a63099c482e57baa7e7016 ]

This ensures that every time the SetIdentity func is called, it will
properly release the old reference in case one is set. This is also the
case when an endpoint is torn down, since we then want to ensure that we
only release the identity if it was actually set.

Signed-off-by: Odin Ugedal <ougedal@ugedal.com>
Signed-off-by: Odin Ugedal <ougedal@palantir.com>
Signed-off-by: Tobias Klauser <tobias@cilium.io>