endpoint: fix endpoint policy deadlock waiting-to-regenerate
[ upstream commit 88f4032c28cb182da56519f7c80f6594728bbe55 ]

This fixes a deadlock caused by context cancellation of the request context
after the "PUT /endpoint/{id}" endpoint successfully returns. This
results in the endpoint regeneration being cancelled in the middle of
the regeneration, leaving the endpoint stuck in "waiting-to-regenerate".
This results in succeeding regenerations "waiting" for the current one
that is cancelled, and the endpoint never progresses. The only
mitigation is restarting the agent, or deleting the pod.

This is very rare, but can happen on pods where the initial regeneration
by "RunMetadataResolver" is cancelled due to its identity changing AND
its regenerated by other parts of the agent. Then, when the
"ep.Regenerate" call using the request context starts,
ep.WaitForFirstRegeneration will return immediately since the endpoint
is ready to go. Then, there is a small chance that the request ends
before the endpoint moves off "waiting-to-regenerate".

Even if this can happen once every blue moon, the consequences are so
severe that and hard to debug, that its worth fixing.

This is safe because endpoint regeneration is always using the aliveCtx
from the endpoint, and in case the "PUT /endpoint/{id}" call fails - it
will delete the endpoint and trigger the aliveCtx to cancel. This
ensures that there should be no leaks as a result of this.

This has been an issue for many releases, and we have reproduced this on
v1.16 and current main (~v1.19).

Signed-off-by: Odin Ugedal <ougedal@palantir.com>
Signed-off-by: Odin Ugedal <odin@ugedal.com>
Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>