kvstore/allocator: let [RunGC] correctly delete duplicated identities
[ upstream commit c331d94a7bda0336fcf4e61ba6dadb3575fcccc3 ]

Currently, the [RunGC] logic does not delete duplicated identities
(i.e., pointing to the same labels), as long as at least one of them
is actually used. Although duplicated identities should never occur
in kvstore mode thanks to locking, this has been observed happening in
the wild, and it is better to be robust in this case as well. Hence,
let's fine tune the GC logic to consider a given identity as used only
if it has any associated secondary key that also matches the actual ID.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Antony Reynaud <antony.reynaud@isovalent.com>